refactor(car): 자동차 이름에 대한 검증은 모델 내부에서 수행하도록 변경
