+++ src/main/java/racingcar/Program.java
@@ -3,24 +3,22 @@ package racingcar;
 import racingcar.control.input.CarNamesInput;
 import racingcar.control.input.StandardInput;
 import racingcar.control.input.TrialInput;
-import racingcar.control.input.validator.NamesValidator;
 import racingcar.control.input.validator.PositiveValidator;
 import racingcar.control.output.Printable;
 import racingcar.control.output.RacingOutput;
 import racingcar.control.output.RacingWinnersOutput;
-import racingcar.model.Car;
 import racingcar.model.Racing;
 import racingcar.model.dto.CarDto;
 import racingcar.model.dto.RacingDto;
-import racingcar.strategy.RandomMovingStrategy;
 
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
 
+
 public class Program {
     public static void main(String[] args) {
-        StandardInput<List<String>> namesInput = new CarNamesInput(new NamesValidator());
+        StandardInput<List<String>> namesInput = new CarNamesInput();
         StandardInput<Integer> trialInput = new TrialInput(new PositiveValidator());
 
         List<String> names = namesInput.getValue();
