+++ src/main/java/racingcar/control/input/validator/DummyValidator.java
@@ -1,7 +1,5 @@
 package racingcar.control.input.validator;
 
-import racingcar.validator.Validator;
-
 public class DummyValidator<T> implements Validator<T> {
     @Override
     public void verify(T value) {
