+++ src/main/java/racingcar/control/input/validator/Validator.java
@@ -1,0 +1,5 @@
+package racingcar.control.input.validator;
+
+public interface Validator<T> {
+    void verify(T value);
+}
