+++ src/test/java/racingcar/model/CarTest.java
@@ -1,7 +1,10 @@
 package racingcar.model;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -20,4 +23,20 @@ class CarTest {
         assertThat(distance).isEqualTo(movement);
     }
 
+    @Test
+    @DisplayName("이름이 5글자를 초과하면 예외가 나야한다")
+    public void nameLength() {
+        String name = "123456";
+
+        Assertions.assertThatThrownBy(() -> new Car(name, () -> 0))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @ParameterizedTest
+    @NullAndEmptySource
+    @DisplayName("이름이 없거나 null이면 예외가 나야 한다")
+    public void emptyName(String name) {
+        Assertions.assertThatThrownBy(() -> new Car(name, () -> 0))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
 }
